requirements.txt
openvino
optimum-intel
transformers[ja]
onnx
nncf[pytorch,tensorflow]
